Rename to_h to parse_json
